Move test env to `given` subpackage
